refactor(help): move helpJumpDispatcher + helpEntry types to @/shared/help
Greptile flagged a feature→shell import cycle introduced by PR 1744:
ParameterPanel (feature) imported helpJumpEventName from shell, while
the shell-side aggregator already imported helpEntries from
@/features/bin-designer. That's a real circular dependency that the
mode-aware PR could not avoid without this move.

Fix: relocate the dispatcher and entry types to @/shared/help/. The
shell still consumes them (HelpModal, MobileHelpModal, Sidebar) but no
feature imports from shell anymore.

- src/shell/Modals/HelpModal/helpEntry.ts        → src/shared/help/helpEntry.ts
- src/shell/Modals/HelpModal/helpJumpDispatcher.ts → src/shared/help/helpJumpDispatcher.ts
- ...same for both .test files
- All 12 importers updated
- vitest.config.ts: add src/shared/help/** to the DOM project's
  test globs so the dispatcher's MutationObserver / document tests run
  in jsdom (matching the old behavior at the old location).

No behavior change; only file locations.
