+++ src/features/categories/helpEntries.ts
@@ -3,7 +3,7 @@
  * modal search via `helpEntryAggregator`.
  */
 
-import type { FeatureHelpEntry } from '@/shell/Modals/HelpModal/helpEntry';
+import type { FeatureHelpEntry } from '@/shared/help/helpEntry';
 
 export const helpEntries: FeatureHelpEntry[] = [
   {
