+++ src/shared/help/helpEntry.ts
